requirements.txt
numpy
pandas
matplotlib
pillow
opencv-python
scikit-learn
tensorflow
